vuln-scanning-artifact-registry/cloud-function/requirements.txt
functions-framework==3.*
google-cloud-securitycenter
google-cloud-containeranalysis
google-cloud-storage
